+++ actix-http/src/responses/mod.rs
@@ -1,0 +1,11 @@
+//! HTTP response.
+
+mod builder;
+mod head;
+#[allow(clippy::module_inception)]
+mod response;
+
+pub use self::builder::ResponseBuilder;
+pub(crate) use self::head::BoxedResponseHead;
+pub use self::head::ResponseHead;
+pub use self::response::Response;
